build(cli): Move the API to create ORT commands to a separate project
This is going to avoid a circular project dependency when the commands
themselves are also moved to separate projects: Otherwise a command
would need to depend on the cli for the API, but the cli also needs to
depend on the command platform project.
