+++ cli/build.gradle.kts
@@ -163,6 +163,7 @@ dependencies {
     implementation(project(":evaluator"))
     implementation(project(":model"))
     implementation(project(":notifier"))
+    implementation(project(":plugins:commands:command-api"))
     implementation(project(":reporter"))
     implementation(project(":scanner"))
     implementation(project(":utils:ort-utils"))
