+++ cli/src/main/resources/META-INF/services/org.ossreviewtoolkit.plugins.commands.api.OrtCommand
