+++ plugins/commands/api/src/main/kotlin/OrtCommand.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.cli
+package org.ossreviewtoolkit.plugins.commands.api
 
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.UsageError
